Stash selected tab state and fix state stash overwriting
