+++ bundles/framework/layerlist/view/LayerViewTabs/index.js
@@ -1,2 +1,5 @@
 export { LayerViewTabs } from './LayerViewTabs';
 export { LayerViewTabsHandler } from './LayerViewTabsHandler';
+
+export const TABS_ALL_LAYERS = 'ALL';
+export const TABS_SELECTED_LAYERS = 'SELECTED';
